+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fixed a migration error that could occur when updating to Commerce 3. ([#1726](https://github.com/craftcms/commerce/issues/1726)) 
 - Fixed a bug that included control panel styles in rendered email previews. ([#1734](https://github.com/craftcms/commerce/issues/1734))
 - Fixed a bug where  the `countryText` attribute was missing in the `craft\commerce\models\Address:getAddressLines()` return data. 
+- Fixed a PHP error that occurred when saving an order without a customer ID. 
 
 ## 3.2.6 - 2020-09-13